requirements.txt
absl-py==1.1.0
anyio==3.6.1
asgiref==3.5.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
fastapi==0.78.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
gunicorn==20.1.0
h11==0.13.0
h5py==3.7.0
idna==3.3
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-serving-api==2.9.0
termcolor==1.1.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.2
wrapt==1.14.1
